An S4 class object of method-assays

An S4 class object that contains data matrices, HVGs, feature metadata, neighbourhood graphs, reduction embeddings, cluster assignments, and benchmarking results.

Slots

counts

Raw counts matrix in dgCMatrix format.

normalised

Normalised matrix in dgCMatrix format.

norm.scaled

HVG subset of normalised and scaled matrix in matrix format.

highly.variable.genes

list of highly variable genes in character format.

feature_metadata

data frame containing feature-level metadata.

graphs

list of generated neighbourhood graphs (distances)

computational_reductions

list of reductions to improve computational efficiency, i.e. PCA or dbMAP

integration_reductions

list of integration output in reduction format

visualisation_reductions

list of visualisation reductions, i.e. t-SNE, UMAP, etc.

cluster_assignments

list of cluster assignment output. A dataframe is assigned to each method containing subsequent parameter changes.

benchmarking_results

list of benchmarking results corresponding to the cluster assignment dataframes.