Performs graph-based clustering

Performs graph-based clustering on previously generated neighbouhood graphs.

perform.graph.cluster(
  object,
  assay,
  neighbours,
  algorithm = 1,
  cluster.df.name.suffix = "",
  res = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 1.1, 1.2, 1.3, 1.4, 1.5),
  ...
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

neighbours
Character. String indicating which neighbourhood graphs should be used.

algorithm
Numerical. Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python. Default = 1 Default = NULL

res
Numerical vector. Which resolution to run the clusterign algorithm at, a smaller and larger value identified less and more clusters, respectively. Default = c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2,1.3,1.4,1.5)

...
arguments to be passed to Seurat::FindClusters

cluster.df.name
Character. What to call the df contained in clusters. Default = 'seurat

Value

Cluster assignments using the list of resolutions provided contained within cluster_assignments under cluster.df.name