Performs SC3 clustering on reduced embeddings

Performs SC3 clustering on defined method-assays and supplied reductions. SC3 is designed to function on count matrices, reduction may not always work well.

perform.sc3.reduction.cluster(
  object,
  assay,
  reduction,
  dims,
  cluster.df.name.suffix = "",
  ks,
  n.core = 3
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

reduction
Character. String defining which reduction to supply to the clustering algorithm.

dims
Numerical. How many dimensions of the reduciton should be supplied, NULL equates to all.

ks
Numerical range. Number of clusters to identify, this can be a range, i.e. 5:10.

n.core
Numerical. How many cores should be used to run SC3. Default = 3

assignment.df.name
Character. What to call the df contained in clusters.

Value

Cluster assignments using the list of resolutions provided contained within cluster_assignments under assignment.df.name