Performs Scanorama integration

Performs Scanorama integration on defined method-assays and reductions contained within. This is performed on reductions.

perform.scanorama(
  object,
  assay,
  slot = "norm.scaled",
  split.by,
  n.dims = 50,
  reduction.save.suffix = "",
  batch_size = 5000,
  approx = TRUE,
  sigma = 15,
  alpha = 0.1,
  knn = 20,
  union = FALSE,
  seed = 12345
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

slot
Character. String defining which slot in the assay to supply to Scanorama. Default = NULL

split.by
Character. indicating the metadata column containing the batch to split the assay by.

n.dims
Numerical. The number of Scanorama dimensions to be produced. Default = 50

reduction.save.suffix
Character. Should a suffix be added to the end of scanorama, This cannot include underscores.

batch_size
Numerical. The batch size used in the alignment vector computation. Useful when integrating large datasets. Default = 5000

approx
Boolean. Use appoximate nearest neighbours within python, speeds up runtime. Default = TRUE

sigma
Numerical. Correction smoothing parameter on Gaussian kernel. Default = 15

alpha
Numerical. Alignment score minimum cutoff. Default = 0.1

knn
Numerical. Number of nearest neighbors to use for matching. Default = 20

union
Boolean. Should genes between datasets be intersected or not. Default = FALSE

seed
Numerical. The seed to use when integrating these datasets. Default = 12345

Value

Scanorama reduction saved in the supplied method-assays

Examples

object <- perform.scanorama(object = object, 
                            assay = c('SCT', 'SCRAN', 'SCANPY'), 
                            slot = 'norm.scaled', 
                            split.by = 'original.project', 
                            n.dims = 50)
#> Error in is(object = object, class2 = "IBRAP") object 'object' not found