Performs Scanpy normalisation, hvg selection, scaling and variance stabilisation and regression.

A new method-assay is produced. Raw counts are normalised and HVGs identified using Scanpy

perform.scanpy(
  object,
  assay = "RAW",
  slot = "counts",
  new.assay.name = "SCANPY",
  target_sum = 1e+06,
  exclude_highly_expressed = FALSE,
  max_fraction = 0.05,
  key_added = "scanpy_norm_factor",
  log1 = TRUE,
  n_top_genes = 1500,
  max_mean = 6,
  min_mean = 0.0125,
  min_disp = 0.5,
  span = 0.3,
  n_bins = 20,
  flavor = "seurat",
  batch_key = NULL,
  do.scale = TRUE,
  vars.to.regress = NULL,
  n_jobs = NULL,
  zero_center = TRUE,
  max_value = NULL,
  obsm = NULL
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

slot
Character. String indicating which slot within the assay should be sourced

new.assay.name
Character. What should the new assay be called. Default = 'SCRAN'

target_sum
Numerical. What should the data be scaled to. Default = 1e6

exclude_highly_expressed
Boolean. Should highly expressed genes be excluded. Default = FALSE

max_fraction
Numerical. If exclude_highly_expressed=True, consider cells as highly expressed that have more counts than max_fraction of the original total counts in at least one cell. Default = 0.05

key_added
Character. What should the column name be that contains cell scaling factors. Default = 'scanpy_norm_factor'

n_top_genes
Numerical. How many HVGs should be identified. Default = 1500

max_mean
Numerical. If n_top_genes is NULL, this is the maximum mean to determine HVGs. Default = 6

min_mean
Numerical. If n_top_genes is NULL, this is the minimum mean to determine HVGs. Default = 0.0125

min_disp
Numerical. If n_top_genes is NULL, The minimum dispersion that should be presented in a gene for it to be considered highly varaible. Default = 0.5

span
Numerical. The fraction of cells that should be subset for the LOESS fit model. Default = 0.3

n_bins
Numerical. Number of bins to produce when determining HVGs

batch_key
Character. Which column in the metadata identifies the batches of the cells. Default = NULL

do.scale
Boolean. Whether the gene expression should be scaled. Default = TRUE

vars.to.regress
Character. A single or multiple columns of information in the metadata that should be regressed from the dataset. Default = NULL

n_jobs
Numerical. The number of jobs that should be performed in parallel. Default = NULL

max_value
Numerical. Clip to this value after scaling. Default = NULL

obsm
Numerical. Which value to scale data against. Default = NULL

flavour
Character. Choosing which HVG selection method to use when, options: 'seurat', 'cell_ranger', 'seurat_v3'. Default = 'seurat'

zero_centre
Boolean. Should the gene expression be centred arouynd 0. Default = TRUE

Value

Produces a new 'methods' assay containing normalised, scaled and HVGs.