Plot of violin plot of defined features

Plot of violin plot of defined features

# S3 method for vln
plot(
  object,
  assay,
  slot,
  features,
  group.by,
  title = NULL,
  xlab = "group",
  ylab = "expression"
)

Arguments

object
An IBRAP S4 class object

assay
Character. Which assay within the object to access

slot
Character. Which expression matrix would you like to access. Default = 'normalised'

features
Character. Which features should be plotted, it is not recommended to exceed 3 but can be more if your screen size is larger.

group.by
Vector. What should the cell types be split by.

title
Character. What should be the overall title of the plot. Default = NULL

xlab
Character. What should by the x axis title. Default = 'group'

ylab
Character. What shoul be the y axis title. Default = 'expression'

Value

A violin plot of defined gene expression and groups