Performs BBKNN integration

Performs BBKNN integration on defined method-assays and reductions contained within. This is performed on reductions.

perform.bbknn(
  object,
  assay,
  reduction,
  graph.name.suffix = "",
  batch,
  approx = FALSE,
  metric = "euclidean",
  neighbors_within_batch = 3,
  n_pcs = NULL,
  trim = NULL,
  n_trees = 10,
  use_faiss = TRUE,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  generate.diffmap = FALSE,
  n_comps = 15,
  diffmap.name.suffix = ""
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

reduction
Character. String defining the name of the reduction to provide for BBKNN. Default = NULL

graph.name.suffix
Character. Should a suffix be added to the end of bbknn as the graph name, i.e. parameter changes?

batch
Character. Column name in metadata indicating batch. Can be multiple.

approx
Character. Employs annoy's approximate neighbour finding. Useful for large datasets but may increase correction.

neighbors_within_batch
Numerical. How many neighbours to report per batch. Default = 3

n_pcs
Numerical. Range of principal components to use. Default = NULL

trim
Numerical. Trims the n of neighbours per cell to this value. Helps with population independence. Default = NULL

n_trees
Numerical. Number of trees to generate in annoy forest. More trees provides higher precision at the cost of increased resource demand and run time. Default = 10

use_faiss
Boolean. Uses faiss package to compute nearest neighbour, this improves run time at the cost of precision. Default = TRUE

set_op_mix_ratio
Numerical. UMAP connectivity parameter between 0 and 1. controls the blen d between a connectivity matrix formed exclusively from mutual nearest neighbour pairs (0) and a union of all observed neighbour relationships with the mutual pairs emphasised (1). Default = 1.0

local_connectivity
Numerical. How many nearest neighbours of each cell are assumed to be fully connected. Default = 1

generate.diffmap
Boolean. Should diffusion maps be generated from the neighourhood graphs, these will be stored in computational_reductions and can be used for umap generation and further neighbourhood generation. Default = TRUE

n_comps
Numerical. How many components should be generated for the diffusion maps. Default = 15

metric.
Character. Which distance metric to use when approx is TRUE, options: 'angular', 'euclidean', 'manhattan' or 'hamming'. Default = 'euclidean'

diffmap.name.sufix
Character. Should a suffix be added to the end of bbknn:diffmap as the reduction name, i.e. parameter changes?

Value

BBKNN connectivity graph contained in graphs in the indicated method-assays