Performs Harmony integration

Performs Harmony integration on defined method-assays and reductions contained within. This is performed on reductions.

perform.harmony(
  object,
  assay,
  vars.use,
  reduction = "pca",
  reduction.save.suffix = "",
  dims.use = NULL,
  save.plot = TRUE,
  ...
)

Arguments

object
IBRAP S4 class object

assay
Character. String containing indicating which assay to use

vars.use
Character. A string of the column nmae that contains variables to regress.

reduction
Character. String defining the name of the reduction to provide for BBKNN. Default = 'pca'

reduction.save.suffix
Character. What should be appended to the end of harmony as the reduction name

dims.use
Numerical. Number of dimensions of the provided reduction to input into harmony, NULL equates to all dimensions. Default = NULL

save.plot
Boolean. Should the automatically genewrated plot be saved? Default = TRUE

...
Arguments to be passed to harmony::HarmonyMatrix

Value

PCA reductions contained within the computational_reduction list in the defined assays

Examples

object <- perform.harmony(object = object, 
                          assay = c('SCRAN', 'SCT', 'SCANPY'), 
                          vars.use = 'original.project', 
                          reduction = c('pca'),  
                          max.iter.harmony = 100,
                          dims.use = list(NULL))
#> Error in is(object = object, class2 = "IBRAP") object 'object' not found